pyproject.toml
[tool.poetry]
name = "rollerbot"
version = "0.1.0"
description = "A discord bot to roll dice"
authors = ["Nick Thurmes <nthurmes@gmail.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.6"
"discord.py" = "^1.2"
dndice = "^2.4"

[tool.poetry.dev-dependencies]
nose2 = "^0.9.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
